Remove AdapterActionFuture
With the removal of `AdapterActionFuture#convert` in #93140 there are
now no differences between `PlainActionFuture` and
`AdapterActionFuture`, so this commit combines them.
