+++ server/src/main/java/org/elasticsearch/action/support/ListenableActionFuture.java
@@ -17,7 +17,7 @@ import java.util.List;
  * A {@code Future} and {@link ActionListener} against which other {@link ActionListener}s can be registered later, to support
  * fanning-out a result to a dynamic collection of listeners.
  */
-public class ListenableActionFuture<T> extends AdapterActionFuture<T> {
+public class ListenableActionFuture<T> extends PlainActionFuture<T> {
 
     private Object listeners;
     private boolean executedListeners = false;
